refactor(github): extract SWR cache and search logic into useGitHubResourceListSWR hook
- Move fetchData, mount/load effect, focus-revalidate effect, and numeric query effect from 1383-line component into co-located hook
- Hook owns all data/loading/error state and returns stable handleLoadMore/handleRetry/handleManualRefresh callbacks
- Preserve all SWR comment blocks documenting warm-slot hydration paths
- Keep generation-token + cache-key invariant (lastLoadedEffectKeyRef) in single module
